NO-TICKET: Move genesis-related structs and functions to separate file.
