+++ execution-engine/comm/src/engine_server/mod.rs
@@ -1,8 +1,3 @@
-pub mod ipc;
-pub mod ipc_grpc;
-pub mod mappings;
-pub mod state;
-
 use std::convert::TryInto;
 use std::fmt::Debug;
 use std::io::ErrorKind;
@@ -10,22 +5,27 @@ use std::marker::{Send, Sync};
 use std::time::Instant;
 
 use common::key::Key;
+use common::value::account::BlockTime;
 use common::value::U512;
+use execution_engine::engine_state::{EngineState, genesis::GenesisResult};
 use execution_engine::engine_state::error::Error as EngineError;
 use execution_engine::engine_state::execution_result::ExecutionResult;
-use execution_engine::engine_state::{EngineState, GenesisResult};
 use execution_engine::execution::{Executor, WasmiExecutor};
 use execution_engine::tracking_copy::QueryResult;
 use shared::logging;
 use shared::logging::{log_duration, log_info};
 use shared::newtypes::{Blake2bHash, CorrelationId};
 use storage::global_state::History;
-use wasm_prep::wasm_costs::WasmCosts;
 use wasm_prep::{Preprocessor, WasmiPreprocessor};
+use wasm_prep::wasm_costs::WasmCosts;
 
 use self::ipc_grpc::ExecutionEngineService;
 use self::mappings::*;
-use common::value::account::BlockTime;
+
+pub mod ipc;
+pub mod ipc_grpc;
+pub mod mappings;
+pub mod state;
 
 const EXPECTED_PUBLIC_KEY_LENGTH: usize = 32;
